requirements.txt
nb-cli
nonebot-adapter-cqhttp
nonebot_plugin_apscheduler
psutil
numpy
redis